requirements.txt
python-dotenv
requests
Jinja2